Add files via upload
dodany screen i fizyka

+++ src/Entity/Bullet.java
@@ -3,8 +3,8 @@ package Entity;
 public class Bullet extends Entity {
     private boolean isAlly;
 
-    public Bullet(int hp, byte radius, float angle,
-                  int[] position, int movementSpeed, boolean isAlly) {
+    public Bullet(int hp, float radius, float angle,
+                  float[] position, int movementSpeed, boolean isAlly) {
         super(hp, radius, angle, position, movementSpeed);
         this.isAlly = isAlly;
     }
